style: added deletion confirmation for list

+++ src/app/components/user-list/user-list.component.ts
@@ -1,16 +1,21 @@
-import {Component, OnInit} from '@angular/core';
+import {Component, OnDestroy, OnInit} from '@angular/core';
 import {User, UsersService} from '../../services/users/users.service';
+import {ConfirmDialogService} from '../confirm-dialog/confirm-dialog.service';
+import {Subject, takeUntil} from 'rxjs';
 
 @Component({
   selector: 'user-list',
   templateUrl: './user-list.component.html',
   styleUrl: './user-list.component.scss'
 })
-export class UserListComponent implements OnInit {
+export class UserListComponent implements OnInit, OnDestroy {
   users!: User[];
 
+  destroy: Subject<void> = new Subject<void>();
+
   constructor(
-    private readonly usersService: UsersService
+    private readonly usersService: UsersService,
+    private readonly confirmDialogService: ConfirmDialogService
   ) {
   }
 
@@ -22,6 +27,14 @@ export class UserListComponent implements OnInit {
   }
 
   removeUser(id: number): void {
-    this.usersService.remove(id);
+    this.confirmDialogService.openConfirmationDialog()
+      .pipe(takeUntil(this.destroy))
+      .subscribe(() => {
+        this.usersService.remove(id);
+      });
+  }
+
+  ngOnDestroy(): void {
+    this.destroy.next();
   }
 }
